requirements.txt
Django==2.0.4
djangorestframework==3.8.2
ez-setup==0.9
mysqlclient==1.3.12
Pygments==2.2.0
PyMySQL==0.8.0
pytz==2018.4
